Fix websocket close code not getting sent
When rejecting the websocket handshake, the server just responds with a `404` HTTP status code instead of the `101`.
The browser just calls `onclose` with a closecode of `1006` as if the server was offline, so we don't know if we need to just reconnect or grab a new token.

Wait until the handshake completes and validate the token when the websocket connection is open. If the token is invalid close the connection right away again.